requirements.txt
absl-py==0.9.0
apache-beam==2.25.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
audioread==2.1.9
avro-python3==1.9.2.1
backcall==0.2.0
black==20.8b1
bleach==3.2.1
bokeh==2.2.3
bz2file==0.98
cachetools==4.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
crcmod==1.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
Django==2.2
django-cors-headers==3.2.0
djangorestframework==3.11.0
dm-sonnet==2.0.0
dm-tree==0.1.5
docopt==0.6.2
dopamine-rl==3.0.1
entrypoints==0.3
fastavro==1.2.1
Flask==1.1.2
future==0.18.2
gast==0.2.2
gevent==20.9.0
gin-config==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.34.0
gunicorn==20.0.4
gym==0.17.3
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.8
imageio==2.9.0
importlib-metadata==3.1.1
importlib-resources==3.3.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
-e git+https://github.com/tensorflow/magenta.git@ce51591cde5c3e9f9d49a0cb891fde5b7ec7d48d#egg=magenta
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mesh-tensorflow==0.1.17
mido==1.2.6
mir-eval==0.6
mistune==0.8.4
mock==2.0.0
more-itertools==8.6.0
mpmath==1.1.0
music21==6.3.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
note-seq==0.0.2
notebook==6.1.5
numba==0.48.0
numpy==1.18.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.7
pandas==0.25.3
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
pretty-midi==0.2.9
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
py==1.9.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pydub==0.24.1
pyglet==1.5.0
Pygments==2.7.3
pygtrie==2.4.1
pylint==2.6.0
pymongo==3.11.2
pymusicxml==0.3.6
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.17.3
pytest==6.1.2
pytest-pylint==0.18.0
python-dateutil==2.8.1
python-ly==0.9.6
python-rtmidi==1.1.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.22
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
sk-video==1.1.10
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
sox==1.4.1
sqlparse==0.3.0
sympy==1.7
tabulate==0.8.7
tensor2tensor==1.15.7
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.0.0
tensorflow-addons==0.11.2
tensorflow-datasets==4.1.0
tensorflow-estimator==2.0.1
tensorflow-gan==2.0.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-slim==1.1.0
threadpoolctl==2.1.0
tifffile==2020.12.4
toml==0.10.2
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typed-ast==1.4.1
typeguard==2.10.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
yapf==0.30.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
